backend/pyproject.toml
[project]
name = "embedding-pipeline"
version = "0.1.0"
description = "Embedding pipeline to extract text from Docusaurus URLs, generate embeddings using Cohere, and store in Qdrant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.9.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"